% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawPCAplot.R
\name{drawPCAplot}
\alias{drawPCAplot}
\title{Draw PCA plot after running LorDist}
\usage{
drawPCAplot(dist.mat, sample_information, SubjectID, Label)
}
\arguments{
\item{dist.mat}{output of LorDist}

\item{sample_information}{Metadata included Subject ID, Label.}

\item{SubjectID}{Column name of Subject ID in Metadata.}

\item{Label}{Column name of Label in Metadata.}
}
\value{
a figure as ggplot2 object
}
\description{
Draw PCA plot after running LorDist
}
\examples{
data(simdata)
dist<-LorDist(SimData.mat, SimSample_information, SampleID = 'SampleID',
              SubjectID = "SubjectID", Timepoint = 'Timepoint')
drawPCAplot(dist, SimSample_information, SubjectID = "SubjectID", Label = "Label")



}
