% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample2fd.R
\name{Sample2fd}
\alias{Sample2fd}
\title{Convert data into functional data format}
\usage{
Sample2fd(otu_data, collect_time, starttime, endtime, method, norder = 2)
}
\arguments{
\item{otu_data}{Abundance matrix with data type "data.frame", row is feature, column is sample.}

\item{collect_time}{a vector of sample collection time points}

\item{starttime}{Time to start sample collection}

\item{endtime}{Time to end sample collection}

\item{method}{Curve fitting method. The default is "Bspline", can also choose "Fourier".}

\item{norder}{an integer specifying the order of b-splines. The default of 2.}
}
\value{
an fdobj
}
\description{
Convert data into functional data format
}
\examples{



}
